WIP: trying to get rid of test-runner.sh for 60* tests
The second test fails to setup because it never receives the
NameOwnerChanged signal on the session bus ... I have a feeling this
is because dbus-glib is looking on the wrong bus, I imagine it's not
designed to switch the bus in use at runtime.

Using dbus.bus.BusConnection instead of dbus.SessionBus in
common/utils/helpers.py is probably the way forwards!